feat: Add UserImagePicker widget
This commit adds a new UserImagePicker widget to the project. The widget allows users to select an image from their device and display it in a CircleAvatar. It also includes a button to add a new image. This enhances the user experience by providing a convenient way to choose and display user images.