feat: Update bundle identifier and dependencies for iOS and macOS
This commit updates the bundle identifier in the Xcode project file and the Info.plist file for the iOS Runner target. It also updates the bundle identifier in the Xcode project file for the macOS Runner target. Additionally, it adds the firebase_messaging dependency to the pubspec.yaml file and registers the FLTFirebaseMessagingPlugin in the GeneratedPluginRegistrant.swift file for macOS. These changes ensure that the app uses the correct bundle identifier and includes the necessary dependencies for Firebase Cloud Messaging.